move RotatedShape class out to separate file
